requirements.txt
crispy-bootstrap5
Django==4.2.9
django-crispy-forms
django-environ
whitenoise

